move render methods into their own file
